+++ src/axom/core/utilities/Utilities.hpp
@@ -31,7 +31,6 @@
 
 #include <cassert>         // for assert()
 #include <cmath>           // for log2()
-#include <cstdlib>         // for std::malloc, std::realloc, std::free
 
 #include <random>        // for random  number generator
 #include <type_traits>   // for std::is_floating_point()
@@ -41,55 +40,6 @@ namespace axom
 {
 namespace utilities
 {
-
-/*!
- * \brief Allocates a chunk of memory of type T.
- * \param [in] n the number of elements to allocate.
- * \tparam T the type of pointer returned.
- * \return A pointer to the new allocation or a null pointer if allocation
- *  failed.
- */
-template < typename T >
-inline T* alloc( std::size_t n )
-{
-  return static_cast< T* >( std::malloc( n * sizeof( T ) ) );
-}
-
-/*!
- * \brief Reallocates the chunk of memory pointed to by pointer.
- * \param [in] pointer pointer to memory previously allocated with
- *  alloc or realloc, or a null pointer.
- * \param [in] n the number of elements to allocate.
- * \tparam T the type pointer points to.
- * \return A pointer to the new allocation or a null pointer if allocation
- *  failed.
- */
-template < typename T >
-inline T* realloc( T* pointer, std::size_t n )
-{
-  if ( n == 0 )
-  {
-    std::free( pointer );
-    return nullptr;
-  }
-
-  return static_cast< T* >( std::realloc( pointer,  n * sizeof( T ) ) );
-}
-
-/*!
- * \brief Frees the chunk of memory pointed to by pointer.
- *
- * \param [in] pointer pointer to memory previously allocated with
- *  alloc or realloc or a null pointer.
- *
- *  \post pointer == nullptr
- */
-template < typename T >
-inline void free( T*& pointer )
-{
-  std::free( pointer );
-  pointer = nullptr;
-}
 
 /*!
  * \brief Gracefully aborts the application